happy with current state
Happy with the current state of edge detection. I was unable to get a vertex depth prediction algorithm to perform well. Instead of using this predicted depth to check against the actual depth to get edge I used the difference in the local depth only. Fine tuning the parameters can result in well defined edges. Now off to work on the inside convex edges :)