grass shader v1
grass has an issue where it can take on the colour of objects. This is because it is coloured based on what is behind it. Might be a way to change how the grass is culled. I might be able to set the render queue as well.

Another thing I want to try is using the terrain hight and normals to colour the grass the same way the terrain is coloured